requirements.txt
backcall==0.2.0
boto3==1.18.1
botocore==1.21.65
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colour==0.1.5
cycler==0.12.1
dataclasses==0.6
decorator==5.1.1
docker-pycreds==0.4.0
einops==0.3.2
gitdb==4.0.11
GitPython==3.1.43
idna==3.7
imageio==2.34.1
ipython==7.25.0
jedi==0.19.1
jmespath==0.10.0
joblib==1.4.2
kiwisolver==1.4.5
matplotlib==3.4.2
matplotlib-inline==0.1.7
networkx==3.1
numpy==1.17.4
opencv-python==4.4.0.46
OpenEXR==1.3.8
parso==0.8.4
pexpect==4.9.0
pickleshare==0.7.5
Pillow==9.0.1
platformdirs==4.2.1
prompt-toolkit==3.0.43
protobuf==4.25.3
psutil==5.9.8
ptyprocess==0.7.0
pygments==2.18.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
PyWavelets==1.4.1
PyYAML==6.0.1
requests==2.31.0
s3transfer==0.5.2
scikit-image==0.18.2
scikit-learn==1.1.2
scipy==1.5.4
sentry-sdk==2.1.1
setproctitle==1.3.3
shortuuid==1.0.1
six==1.16.0
smmap==5.0.1
threadpoolctl==3.5.0
tifffile==2023.7.10
timm==0.4.12
# torch==1.10.1+cu111
# torchvision==0.11.2+cu111
tqdm==4.51.0
traitlets==5.14.3
typing-extensions==4.11.0
urllib3==1.26.18
wandb==0.17.0
wcwidth==0.2.13
zstandard==0.15.2
